machines-discord-bot/requirements.txt
discord.py
langchain==0.0.337
faiss-cpu
pinecone.client
python-dotenv
openai
tiktoken
pynacl
pyaudio
pydub
ffmpeg-python
sounddevice
soundfile
pinecone-client==2.2.4
python-docx
cohere
pypdf
docx